# Conflicts:
#	DoAnTotNghiepBE.iml
#	src/main/java/com/example/doantotnghiepbe/entity/ApprovalPost.java
#	src/main/java/com/example/doantotnghiepbe/entity/Post.java
#	src/main/java/com/example/doantotnghiepbe/entity/User.java
#	src/main/java/com/example/doantotnghiepbe/repository/PostRepository.java
#	target/classes/application.properties
#	target/classes/com/example/doantotnghiepbe/entity/ApprovalPost.class
#	target/classes/com/example/doantotnghiepbe/entity/Post.class
#	target/classes/com/example/doantotnghiepbe/entity/User.class
#	target/classes/com/example/doantotnghiepbe/repository/PostRepository.class
#	target/classes/com/example/doantotnghiepbe/repository/UserRepository.class
#	target/classes/com/example/doantotnghiepbe/service/PostDetailService.class
#	target/classes/com/example/doantotnghiepbe/service/impl/CommentServiceImpl.class
#	target/classes/com/example/doantotnghiepbe/service/impl/FavoriteServiceImpl.class
#	target/classes/com/example/doantotnghiepbe/service/impl/PostDetailServiceImpl.class

+++ src/main/java/com/example/doantotnghiepbe/repository/PaymentRepository.java
@@ -8,13 +8,14 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import java.util.Optional;
 
 @Repository
 public interface PaymentRepository extends JpaRepository<Payment, Long> {
     Payment findByVnpTxnRef(String vnpTxnRef);
 
     @Query("SELECT p FROM Payment p WHERE p.postId.user.id = :userId")
-    List<Payment> findPaymentsByUserId(@Param("userId") Integer userId);
+    List<Payment> findPaymentsByUserId(@Param("userId") Long userId);
 
 
 }
